+++ packages/components/atoms/f-image-tile/src/components/ImageTile.vue
@@ -1,6 +1,9 @@
 <template>
     <div
-        :class="$style['c-imageTile']"
+        :class="[
+            $style['c-imageTile'], {
+                [$style['c-imageTile--selected']]: isToggleSelected
+            }]"
         data-test-id="image-tile-component">
         <a
             :class="[
@@ -23,29 +26,40 @@
             data-test-id="image-tile-input"
             @change="toggleFilter">
         <label
-            :class="[
-                $style['c-imageTile-label'], {
-                    [$style['c-imageTile-label--selected']]: isToggleSelected
-                }]"
+            :class="$style['c-imageTile-label']"
             :for="`imageTileToggle-${tileId}`"
             data-test-id="image-tile-label">
-            <img
-                :class="$style['c-imageTile-image']"
-                :src="imgSrc"
-                data-test-id="image-tile-image"
-                alt="">
-            <span :aria-hidden="isLink">
-                {{ displayText }}
+            <span
+                :class="$style['c-imageTile-imageContainer']"
+                :style="cssVars">
+                <img
+                    v-if="imgSrc"
+                    :class="$style['c-imageTile-image']"
+                    :src="imgSrc"
+                    data-test-id="image-tile-image"
+                    :alt="altText"
+                    :role="isPresentationRole ? 'presentation' : false">
+            </span>
+            <span
+                :class="$style['c-imageTile-textContainer']"
+                :aria-hidden="isLink">
+                <tick-icon :class="$style['c-imageTile-icon']" />
+                <span :class="$style['c-imageTile-text']">
+                    {{ displayText }}
+                </span>
             </span>
         </label>
     </div>
 </template>
 
 <script>
+import { TickIcon } from '@justeat/f-vue-icons';
 
 export default {
     name: 'ImageTile',
-    components: {},
+    components: {
+        TickIcon
+    },
     props: {
         href: {
             type: String,
@@ -69,13 +83,37 @@ export default {
         },
         imgSrc: {
             type: String,
-            default: null
+            default: ''
+        },
+        altText: {
+            type: String,
+            default: ''
+        },
+        fallbackImage: {
+            type: String,
+            default: ''
         }
     },
     data () {
         return {
             isToggleSelected: false
         };
+    },
+    computed: {
+        isPresentationRole () {
+            return this.altText === '';
+        },
+        /**
+         * Returns a css variable from the fallback image prop
+         *
+         */
+        cssVars () {
+            const cssVariable = JSON.stringify(this.fallbackImage);
+
+            return {
+                '--bg-image': `url(${cssVariable})`
+            };
+        }
     },
     watch: {
         isSelected (newValue) {
@@ -117,11 +155,22 @@ export default {
 
 <style lang="scss" module>
 
-.c-imageTile {
-    @include font-size(heading-m);
+$image-tile-background-opacity: 0.7;
+$image-tile-background-color: $color-interactive-brand;
+$image-tile-selected: $color-content-positive;
+$image-tile-transition-duration: 0.2s;
+$image-tile-ease: ease-in-out;
+$image-tile-text-transform: translate3d(5px, 0, 0);
 
-    font-family: $font-family-base;
+@mixin image-tile-icon-selected-transform() {
+    opacity: 1;
+    transform: translate3d(0, 0, 0) scale(1) rotate(0);
+    width: 15px;
+}
+
+.c-imageTile {
     position: relative;
+    width: 100%;
 }
 
 .c-imageTile-link {
@@ -143,12 +192,89 @@ export default {
     flex-flow: column wrap;
 }
 
-.c-imageTile-label--selected {
-  border: 1px solid green;
+.c-imageTile-imageContainer {
+    background-color: rgba($image-tile-background-color, $image-tile-background-opacity);
+    border-radius: $radius-rounded-b;
+    background-image: var(--bg-image);
+    display: block;
+    padding-top: (3 / 5) * 100%; // 5:3 aspect ratio
+    position: relative;
 }
 
 .c-imageTile-image {
+    border-radius: $radius-rounded-b;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+}
+
+.c-imageTile-textContainer {
+    margin-top: spacing(x2);
+    display: flex;
+    max-width: 100%;
+
+    .c-imageTile--selected & {
+        color: $image-tile-selected;
+    }
+}
+
+.c-imageTile-icon {
+    align-self: center;
+    opacity: 0;
+    transform: translate3d(-10px, 0, 0) scale(0.5) rotate(-60deg);
+    width: 0;
+    will-change: transform, opacity;
+
+    @media (prefers-reduced-motion: no-preference) {
+        transition: transform $image-tile-transition-duration $image-tile-ease,
+                    opacity $image-tile-transition-duration $image-tile-ease,
+                    width $image-tile-transition-duration $image-tile-ease;
+    }
+
+    .c-imageTile--selected & {
+        @include image-tile-icon-selected-transform();
+    }
+
+    .c-imageTile--selected & path {
+        fill: $image-tile-selected;
+    }
+
+    @include media('>=mid') {
+        .c-imageTile:hover & {
+            @include image-tile-icon-selected-transform();
+        }
+    }
+}
+
+.c-imageTile-text {
+    @include font-size(body-s);
     display: block;
+    font-family: $font-family-base;
+    font-weight: $font-weight-regular;
+    margin-right: spacing(x2);
+    overflow: hidden;
+    text-overflow: ellipsis;
+    transform: translate3d(0, 0, 0);
+    white-space: nowrap;
+    width: 100%;
+    will-change: transform;
+
+    @media (prefers-reduced-motion: no-preference) {
+        transition: transform $image-tile-transition-duration $image-tile-ease;
+    }
+
+    .c-imageTile--selected & {
+        transform: $image-tile-text-transform;
+    }
+
+    @include media('>=mid') {
+        .c-imageTile:hover & {
+            transform: $image-tile-text-transform;
+        }
+    }
 }
 
 </style>
+