+++ packages/components/organisms/f-header/src/tenants/en-AU.js
@@ -29,6 +29,10 @@ export default {
         }
     },
 
+    userMenu: {
+        buttonLabel: name => `${name}'s account`
+    },
+
     logo: {
         gtm: 'click_logo'
     },