+++ packages/components/molecules/f-restaurant-card/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.14.0
+------------------------------
+*January 20, 2022*
+
+### Changed
+- Fallback to gray background for restaurant image when image doesn't load or is missing
+- Hide restaurant logo when image fails to load
+- Add presentation role to restaurant logo
+- Move restaurant logo to it's own folder
+
 v0.13.0
 ------------------------------
 *January 17, 2022*