+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -3,6 +3,29 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v3.15.3
+------------------------------
+*January 21, 2022*
+
+### Changed
+- Passed tenant to `getCustomer` call
+  
+
+v3.15.2
+------------------------------
+*January 20, 2022*
+
+### Changed
+- Updated version of `f-mega-modal`.
+
+
+v3.15.1
+------------------------------
+*January 20, 2022*
+
+### Fixed
+- Correct key for delivery note type when split notes is not enabled
+
 
 v3.15.0
 ------------------------------